serializationclarity object serialization using our own impl of serialize(), deserialize using specific JSON parsing struct
